refactor(lyrics): import spotify via its public surface, not internal paths
lyrics reached into spotify's internals — '@flows/spotify/src/backend/database'
and '.../src/backend/repository' — coupling it to spotify's file layout.
Both symbols (musicDb, SQLiteTrackRepository) are already exported from
spotify's public index, so import from '@flows/spotify' instead.

Verified: tsc --noEmit passes for spotify, lyrics, trading, chat, and
svelte-check passes for ui.
